pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "transaction-system"
version = "0.1.0"
description = "This is a simple Django REST API for transaction system"
authors = ["temirovazat <azatbekt5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework-jsonapi = {extras = ["django-filter"], version = "^6.1.0"}
mysqlclient = "^2.2.0"
flake8 = "^6.1.0"
black = "^23.9.1"
isort = "^5.12.0"
bandit = "^1.7.5"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
envparse = "^0.2.0"

[tool.bandit]
skips = [
    "B101",  # assert_used
    "B404",  # blacklist_subprocess
    "B603",  # subprocess_without_shell_equals_true
    "B607",  # start_process_with_partial_path
]


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
